requirements.txt
lxml==4.3.0
redis==3.0.1
Django==2.2.3
requests==2.21.0
cassandra_driver==3.18.0
